Fix build error with seed (#25)

+++ candle-core/src/metal_backend/mod.rs
@@ -2106,7 +2106,7 @@ impl BackendDevice for MetalDevice {
         }
         seed_buffer.did_modify_range(metal::NSRange::new(0, 4));
 
-        *self.seed_value.write().unwrap() = seed;
+        *self.seed_value.write().unwrap() = seed as u64;
 
         Ok(())
     }
